requirements.txt
torch
torchvision
tqdm
Pillow
pyyaml
onnx
onnxruntime
albumentations